api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
twitter-api-client>=0.10.22
langchain>=0.1.0
python-dotenv>=0.19.0
pydantic>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
emoji>=2.2.0
nltk>=3.8.1
openai>=1.12.0
transformers>=4.40.0 
sentence-transformers>=3.0.0  
mem0ai
tqdm>=4.66.1
einops>=0.7.0
httpx>=0.24.0 
python-multipart